feat(*)!: Updates pkg-loader to use oci-wasm crate
This adds support for the new oci-wasm crate and uses the types and
client wrapper from that crate. This is marked as a breaking change
because it will only support the new layer content type of
`application/wasm` rather than the older format. I figured that would be
fine since this is a new crate.

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>