feat(wkg)!: Adds support for pushing and pulling OCI wasm artifacts
This adds support for pushing and pulling compatible OCI wasm artifacts
from the command line. As part of this refactor, it makes the original
loading feature a separate command for use in debugging. Some of the
features from the `load` subcommand should probably make their way into
the push and pull command, but for now just keeping it simple

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>