feat(caching): Adds a new caching client
This adds a trait so it can be extendible and provides a default
filesystem implementation. I chose this way of doing it because I didn't
think the hassle of having a `Client<T>` was worth it just to add caching.
This is set up in such a way that you can still use a caching client as a
normal client

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>