feat(config): Adds `Clone` to the `Config` type
We take an owned config to create a client, which is probably the correct
design. Because of this, we should probably have it clonable so it is
easier to use.

I did run into this when adding this to cargo component. We might also
want to make `Client` clonable down the line, but for now this should
work

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>