fix(common): Fixes issue where we couldn't get a backend type
This fixes an issue where if you have a single backend type set, you
still would get a None backend type. This meant that unless you explicitly
specified a type, it would always resolve to OCI as used by the client.

This takes a middle ground approach that lets people write less toml. If
an explicit type is not specified and only one type of config exists for
a registry, then the backend type of that single config will be used
instead. For more than one, you must specify a default.

Please note that I renamed this from `type` to `default` as that term
better expresses the intent of what this config is doing. A user isn't
specifying the type of config so much as the default backend config to
use here in the (currently) unlikely case you have a registry that talks
more than one protocol

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>