fix: Removes use of dirs in favor of etcetera
The `dirs` crate doesn't actually respect XDG vars, so this swaps out to
etcetera. This also has the side effect of removing a transitive dep that
had an MPL licenses (which can cause headaches for CNCF and other
projects).

Also adds a global path for the file cache

Fixes #111
Fixes #98

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>